+++ R/mspeaklists.R
@@ -30,6 +30,13 @@ NULL
 #' @templateVar selj feature groups
 #' @templateVar selOrderj groupNames()
 #' @templateVar optionalji TRUE
+#' @templateVar del TRUE
+#' @templateVar deli feature groups
+#' @templateVar delj MS peaks
+#' @templateVar deljtype numeric indices (rows)
+#' @templateVar delfwhat feature group
+#' @templateVar delfa the peak list table (a \code{data.table}), feature group name, analysis (\code{NULL} if \code{k=NULL}), type (\code{"MS"} or \code{"MSMS"})
+#' @templateVar delfr the peak list indices (rows) to be removed (specified as an \code{integer} or \code{logical} vector)
 #' @templateVar dollarOpName feature group
 #' @template sub_sel_del-args
 #'
@@ -363,6 +370,100 @@ setMethod("as.data.table", "MSPeakLists", function(x, fGroups = NULL, averaged =
     }
 
     return(ret)
+})
+
+#' @templateVar where MSPeakLists
+#' @templateVar what peaks from MS peak lists
+#' @template delete
+#' @export
+setMethod("delete", "MSPeakLists", function(obj, i = NULL, j = NULL, k = NULL, reAverage = FALSE, ...)
+{
+    # NOTE: this is ~ a c/p from the featureAnnotations method
+
+    doAnaOnly <- !is.null(k)
+    
+    ac <- checkmate::makeAssertCollection()
+    i <- assertDeleteArgAndToChr(i, groupNames(obj), add = ac)
+    checkmate::assert(
+        checkmate::checkFunction(j, null.ok = TRUE),
+        checkmate::checkIntegerish(j, null.ok = TRUE),
+        .var.name = "j"
+    )
+    k <- assertDeleteArgAndToChr(k, analyses(obj), add = ac)
+    checkmate::assertFlag(reAverage, add = ac)
+    checkmate::reportAssertions(ac)
+    
+    if ((length(i) == 0 && length(k) == 0) || (!is.null(j) && length(j) == 0))
+        return(obj) # nothing to remove...
+
+    # i/j = NULL; k = vector: remove analyses (subset) (averagedPeakLists ignored)
+    # i != NULL; j = NULL; k = vector: remove groups from specified analyses (averagedPeakLists ignored)
+    # i = NULL; j != NULL; k = vector: remove from all groups from specified analyses (averagedPeakLists ignored)
+    # i != NULL; j != NULL; k = vector: remove from specified groups from specified analyses (averagedPeakLists ignored)
+    # i = vector; j = NULL; k = NULL: remove specified groups (peakLists and averagedPeakLists)
+    # i = vector; j != NULL; k = NULL: remove from specified groups (peakLists and averagedPeakLists)
+    # i = NULL; j != NULL; k = NULL: remove from all groups (peakLists and averagedPeakLists)
+
+    if (!is.null(j) && !is.function(j))
+    {
+        origj <- j
+        j <- function(pl, ...) origj
+    }
+    
+    doDelPLs <- function(PL, grp, ana, ...)
+    {
+        for (t in c("MS", "MSMS"))
+        {
+            if (!is.null(PL[[t]]))
+            {
+                rm <- j(PL[[t]], grp, ana, t, ...)
+                PL[[t]] <- if (is.logical(rm))
+                    PL[[t]][!rm]
+                else
+                    PL[[t]][setdiff(seq_len(nrow(PL[[t]])), rm)]
+            }
+        }
+        return(pruneList(PL, checkZeroRows = TRUE))
+    }
+    
+    if (doAnaOnly && setequal(i, groupNames(obj)) && is.null(j))
+    {
+        # analyses subset
+        obj@peakLists <- pruneList(obj@peakLists[setdiff(analyses(obj), k)])
+    }
+    else
+    {
+        if (is.null(j))
+        {
+            # remove results from groups completely
+            keepFG <- setdiff(groupNames(obj), i)
+            obj@peakLists <- lapply(obj@peakLists, function(x) return(pruneList(x[keepFG])))
+            obj@peakLists <- pruneList(obj@peakLists, TRUE)
+            
+            if (!doAnaOnly)
+                obj@averagedPeakLists <- pruneList(obj@averagedPeakLists[keepFG], TRUE)
+        }
+        else
+        {
+            obj@peakLists[k] <- Map(obj@peakLists[k], k, f = function(x, ana)
+            {
+                x[i] <- Map(x[i], i, f = doDelPLs, MoreArgs = list(ana, ...))
+                return(pruneList(x, checkEmptyElements = TRUE))
+            })
+
+            if (!doAnaOnly)
+            {
+                obj@averagedPeakLists[i] <- Map(obj@averagedPeakLists[i], i, f = doDelPLs, MoreArgs = list(ana = NULL, ...))
+                obj@averagedPeakLists <- pruneList(obj@averagedPeakLists, checkEmptyElements = TRUE)
+            }
+        }
+        
+    }
+
+    if (reAverage)
+        obj@averagedPeakLists <- averageMSPeakLists(obj)
+    
+    return(obj)
 })
 
 #' @describeIn MSPeakLists provides post filtering of generated MS peak lists, which may further enhance quality of